backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6

# RAG and LangChain
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-text-splitters==0.0.1

# Google Gemini
langchain-google-genai==0.0.6
google-generativeai==0.8.3

# OpenAI (for OpenRouter compatibility)
langchain-openai==0.0.2
openai==1.6.1

# Groq
langchain-groq==0.0.1
groq==0.4.1

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF handling and OCR
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
easyocr==1.7.1
opencv-python==4.8.1.78

# Utilities
numpy==1.24.3
typing-extensions==4.9.0

# Content Moderation
detoxify==0.5.1